pyproject.toml
[tool.poetry]
name = "aegis-registry"
version = "0.1.0"
description = ""
authors = ["ikigami0513 <ikigami0513@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
python-dotenv = "^1.2.1"
django-unfold = "^0.73.1"
markdown = "^3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
